+++ src/assignments/solver.py
@@ -1,6 +1,6 @@
 from src.model.group.group import Group
 from src.model.tutor.tutor import Tutor
-from src.model.formatter.output.output_formatter import OutputFormatter
+from src.io.output.output_formatter import OutputFormatter
 
 
 class Solver:
