+++ src/assignments/topic_tutor/group_tutor_solver.py
@@ -2,7 +2,7 @@ from solver import Solver
 from src.model.group.group import Group
 from src.model.tutor.tutor import Tutor
 from src.model.utils.topic import Topic
-from src.model.formatter.output.output_formatter import OutputFormatter
+from src.io.output.output_formatter import OutputFormatter
 
 
 class GroupTutorSolver(Solver):
