+++ src/io/output/output_formatter.py
@@ -1,6 +1,6 @@
 from typing import Union
-from src.model.formatter.output.flow_formatter import FlowOutputFormatter
-from src.model.formatter.output.lp_formatter import LPOutputFormatter
+from src.io.output.flow_formatter import FlowOutputFormatter
+from src.io.output.lp_formatter import LPOutputFormatter
 from src.exceptions import ResultFormatNotFound
 from src.model.group.group import Group
 from src.model.utils.result import AssignmentResult
