+++ tests/io/output/output_formatter_test.py
@@ -1,7 +1,7 @@
 import pytest
-from src.model.formatter.output.output_formatter import OutputFormatter
-from src.model.formatter.output.flow_formatter import FlowOutputFormatter
-from src.model.formatter.output.lp_formatter import LPOutputFormatter
+from src.io.output.output_formatter import OutputFormatter
+from src.io.output.flow_formatter import FlowOutputFormatter
+from src.io.output.lp_formatter import LPOutputFormatter
 from src.exceptions import ResultFormatNotFound
 from src.model.group.group import Group
 from src.model.utils.result import AssignmentResult
